% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teams.R
\name{add_team_member}
\alias{add_team_member}
\title{Add member to an organisation team}
\usage{
add_team_member(org, team, user)
}
\arguments{
\item{org}{String. Name of GitHub organisation}

\item{team}{String. Name of team}

\item{user}{String. GitHub username of member.}
}
\value{
Logical (invisibly).
}
\description{
Add member to an organisation team
}
\examples{
\dontrun{add_team_member("scotgovanalysis", "test-team", "alice-hannah")}
}
