% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/users.R
\name{check_member}
\alias{check_member}
\title{Check if GitHub user is an organisation member}
\usage{
check_member(org, user)
}
\arguments{
\item{org}{String. Name of GitHub organisation}

\item{user}{String. GitHub user name.}
}
\value{
Logical.
}
\description{
Check if GitHub user is an organisation member
}
\examples{
\dontrun{check_member("scotgovanalysis", "alice-hannah")}
}
