% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/issues.R
\name{close_issue}
\alias{close_issue}
\title{Close issue}
\usage{
close_issue(owner, repo, issue_number)
}
\arguments{
\item{owner}{String. Repository owner (e.g. organisation or user profile).}

\item{repo}{Character vector of repository names.}

\item{issue_number}{Number of issue to close}
}
\value{
A tibble.
}
\description{
Close issue
}
\examples{
\dontrun{close_issue("scotgovanalysis", "ghadmin", 1)}
}
