% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/issues.R
\name{create_review_issue}
\alias{create_review_issue}
\alias{create_issue_comment}
\title{Create a new issue}
\usage{
create_review_issue(owner, repo, assign_user, body, label)

create_issue_comment(owner, repo, issue_number, body)
}
\arguments{
\item{owner}{String. Repository owner (e.g. organisation or user profile).}

\item{repo}{Character vector of repository names.}

\item{assign_user}{A character vector of usernames to assign the issue to.}

\item{body}{A character string to use for the body of the issue.}

\item{label}{A character vector of labels to add to the issue.}

\item{issue_number}{Number of issue to add a comment to.}
}
\value{
API response (invisibly). If the action has been successful, a
success message is printed to the console.
}
\description{
Create a new issue
}
