% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/issue_tasks.R
\name{expand_tasks}
\alias{expand_tasks}
\title{Expand tasks in issue body}
\usage{
expand_tasks(issue_tibble)
}
\arguments{
\item{issue_tibble}{Tibble containing a variable called \code{body}.}
}
\value{
The supplied tibble is returned with two additional character
variables; \code{status} ('complete' or 'incomplete') and \code{task_name}.
}
\description{
Expand tasks in issue body
}
\examples{
issue_tibble <- dplyr::tibble(
   issue_number = c(1, 2),
   body = c("Tasks:\n- [x] task 1\n- [ ] task 2",
            "Tasks:\n- [x] task 1\n- [x] task 2")
)
expand_tasks(issue_tibble)
}
