% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ghadmin-package.R
\docType{package}
\name{ghadmin-package}
\alias{ghadmin}
\alias{ghadmin-package}
\title{ghadmin: GitHub Organisation Administration}
\description{
Code to assist with the administration of GitHub organisations.
}
\author{
\strong{Maintainer}: Alice Hannah \email{alice.hannah@gov.scot}

}
\keyword{internal}
