% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/issues.R
\name{issues}
\alias{issues}
\title{Repository issues}
\usage{
issues(owner, repo, state = c("open", "closed", "all"), labels = NULL)
}
\arguments{
\item{owner}{String. Repository owner (e.g. organisation or user profile).}

\item{repo}{Character vector of repository names.}

\item{state}{Optional. To filter issues returned based on status. Either
"open" (default), "closed" or "all".}

\item{labels}{Optional. Character vector of labels to filter issues by.}
}
\value{
A tibble.
}
\description{
Repository issues
}
\examples{
\dontrun{issues("scotgovanalysis", "ghadmin")}
}
