% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{knit_rmd}
\alias{knit_rmd}
\title{Knit Rmarkdown file}
\usage{
knit_rmd(rmd_path, params)
}
\arguments{
\item{rmd_path}{Character string of file path to Rmarkdown file.}

\item{params}{List of parameters to pass to Rmd file.}
}
\value{
Character string of Rmarkdown content.
}
\description{
Knit Rmarkdown file
}
\examples{
\dontrun{
knit_rmd(here("confirm-membership.Rmd"), list(date = "1 March 2025"))
}
}
