% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/issues.R
\name{member_review_issue}
\alias{member_review_issue}
\alias{member_review_reminder}
\title{Create a new issue for membership review}
\usage{
member_review_issue(owner, repo, assign_user, body_template, deadline, label)

member_review_reminder(owner, repo, issue_number, body_template, deadline)
}
\arguments{
\item{owner}{String. Repository owner (e.g. organisation or user profile).}

\item{repo}{Character vector of repository names.}

\item{assign_user}{A character string of the username to assign the issue to.}

\item{body_template}{A character string of the file path to Rmd template.}

\item{deadline}{A character string to include as deadline to respond; e.g.
"Friday 28 February".}

\item{label}{A character vector of labels to add to the issue.}

\item{issue_number}{Number of issue to add a comment to.}
}
\value{
If the API call is successful, a tibble containing \code{owner}, \code{repo},
\code{user}, \code{issue_number} and for \code{member_review_reminder}, \code{reminder_issued}.
}
\description{
Create a new issue for membership review
}
