% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tasks.R
\name{n_tasks}
\alias{n_tasks}
\alias{n_complete}
\alias{n_incomplete}
\title{Summary of task completion}
\usage{
n_tasks(body)

n_complete(body)

n_incomplete(body)
}
\arguments{
\item{body}{Character string, length 1.}
}
\value{
Integer.
}
\description{
Summary of task completion
}
\details{
Within \code{body}, tasks are identified by square brackets
containing either an 'x' (complete) or a space (incomplete).
}
\examples{
n_complete("Some info: [x] task 1 [ ] task 2")
}
