% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{parse_md}
\alias{parse_md}
\title{Parse markdown file}
\usage{
parse_md(md_path)
}
\arguments{
\item{md_path}{Character. File path of markdown file.}
}
\value{
Character string of markdown content.
}
\description{
Parse markdown file
}
\examples{
\dontrun{parse_md(here("confirm-membership.md"))}
}
