% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repos.R
\name{repo_access}
\alias{repo_access}
\alias{repo_contrib}
\title{Repository user access and contributions}
\usage{
repo_access(owner, repo)

repo_contrib(owner, repo)
}
\arguments{
\item{owner}{String. Repository owner (e.g. organisation or user profile).}

\item{repo}{Character vector of repository names.}
}
\value{
A tibble. For \code{repo_access} a row for all users with access to the
repository. For \code{repo_contrib} a row for all users who have contributed to
the repository.
}
\description{
Repository user access and contributions
}
\examples{
\dontrun{
 repo_access("scotgovanalysis", "welcome")
 repo_access("scotgovanalysis", c("welcome", "sgplot"))

 repo_contrib("scotgovanalysis", "welcome")
 repo_contrib("scotgovanalysis", c("welcome", "sgplot"))
}
}
