% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repos.R
\name{repos}
\alias{repos}
\title{Organisation repositories}
\usage{
repos(org)
}
\arguments{
\item{org}{String. Name of GitHub organisation}
}
\value{
A tibble.
}
\description{
Organisation repositories
}
\examples{
\dontrun{repos("scotgovanalysis")}
}
