% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/issue_tasks.R
\name{task_status}
\alias{task_status}
\title{Summary of task completion}
\usage{
task_status(body)
}
\arguments{
\item{body}{Character string, length 1.}
}
\value{
Integer.
}
\description{
Summary of task completion
}
\details{
Within \code{body}, tasks are identified by square brackets
containing either an 'x' (complete) or a space (incomplete).
}
\examples{
task_status("Some info: [x] task 1 [ ] task 2")
}
