% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teams.R
\name{team_members}
\alias{team_members}
\title{Team members}
\usage{
team_members(org, team)
}
\arguments{
\item{org}{String. Name of GitHub organisation}

\item{team}{String. Name of team}
}
\value{
A tibble.
}
\description{
Team members
}
\examples{
\dontrun{team_members("ScotGovAnalysis", "team_name")}
}
