% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teams.R
\name{team_repos}
\alias{team_repos}
\title{Team repositories}
\usage{
team_repos(org, team)
}
\arguments{
\item{org}{String. Name of GitHub organisation}

\item{team}{String. Name of team}
}
\value{
A tibble.
}
\description{
Team repositories
}
\examples{
\dontrun{team_repos("ScotGovAnalysis", "team_name")}
}
