% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teams.R
\name{teams}
\alias{teams}
\title{Organisation teams}
\usage{
teams(org)
}
\arguments{
\item{org}{String. Name of GitHub organisation}
}
\value{
A tibble.
}
\description{
Organisation teams
}
\examples{
\dontrun{teams("ScotGovAnalysis")}
}
