% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/users.R
\name{users}
\alias{users}
\title{Organisation users}
\usage{
users(org, user_type = c("all", "members", "outside_collaborators"))
}
\arguments{
\item{org}{String. Name of GitHub organisation}

\item{user_type}{One of "all", "member", or "outside" (for outside
collaborators). Defauls to "all".}
}
\value{
A tibble.
}
\description{
Organisation users
}
\examples{
\dontrun{
users("ScotGovAnalysis")
}
}
