% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/review_issues.R
\name{validate_review_issues}
\alias{validate_review_issues}
\title{Validate review issues}
\usage{
validate_review_issues(tib, exp_tasks, exp_label)
}
\arguments{
\item{tib}{A tibble containing at least the following variables:
\code{owner}, \code{repo}, \code{user}, \code{issue_number}, \code{date_opened}, \code{state}, \code{body},
\code{labels}, \code{assignees}. This tibble should be created by joining the record
of review issues opened, and the latest data for these issues from GitHub.}

\item{exp_tasks}{Integer of number of tasks expected to be within the issue
body.}

\item{exp_label}{Character string of expected label assigned to GitHub issue.}
}
\value{
The rows of \code{tib} containing errors are returned as a tibble. If
there are no errors, NULL is returned invisibly.
}
\description{
Validate review issues
}
