% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/issues.R
\name{issues}
\alias{issues}
\alias{issue_number}
\title{Repository issues}
\usage{
issues(owner, repo, state = c("open", "closed", "all"), labels = NULL)

issue_number(owner, repo, issue_number, tasks = FALSE)
}
\arguments{
\item{owner}{String. Repository owner (e.g. organisation or user profile).}

\item{repo}{Character vector of repository names.}

\item{state}{Optional. To filter issues returned based on status. Either
"open" (default), "closed" or "all".}

\item{labels}{Optional. Character vector of labels to filter issues by.}

\item{issue_number}{Numeric. Issue number to return.}

\item{tasks}{Logical. If \code{TRUE}, the returned tibble will contain a row for
each task in each issue. If \code{FALSE}, one row per issue will be returned with
no task detail.}
}
\value{
A tibble.
}
\description{
Repository issues
}
\examples{
\dontrun{
issues("scotgovanalysis", "ghadmin")
issue_number("scotgovanalysis", "ghadmin", 1)
}
}
