% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/review_issues.R
\name{new_review_issue}
\alias{new_review_issue}
\alias{review_comment}
\title{Create a new issue for membership review}
\usage{
new_review_issue(owner, repo, assign_user, body_template, label, ...)

review_comment(owner, repo, issue_number, body_template, ...)
}
\arguments{
\item{owner}{String. Repository owner (e.g. organisation or user profile).}

\item{repo}{Character vector of repository names.}

\item{assign_user}{A character string of the username to assign the issue to.}

\item{body_template}{A character string of the file path to Rmd template.}

\item{label}{A character vector of labels to add to the issue.}

\item{...}{Named values to pass to params of \code{body_template} Rmd.}

\item{issue_number}{Number of issue to add a reminder to.}
}
\value{
If the API call is successful, a tibble containing \code{owner}, \code{repo},
\code{user}, \code{issue_number}, \code{date_opened} (for \code{new_review_issue}) and
\code{date_commented} (for \code{review_comment}).
}
\description{
Create a new issue for membership review
}
